+++ es/ae/contract.js
@@ -32,6 +32,7 @@ import * as R from 'ramda'
 import { isBase64 } from '../utils/crypto'
 import ContractCompilerAPI from '../contract/compiler'
 import ContractBase from '../contract'
+import ContractACI from '../contract/aci'
 
 /**
  * Handle contract call error
@@ -107,7 +108,7 @@ async function contractDecodeData (source, fn, callValue, callResult, options) {
  */
 async function contractCallStatic (source, address, name, args = [], { top, options = {} } = {}) {
   const opt = R.merge(this.Ae.defaults, options)
-  const callerId = await this.address(opt)
+  const callerId = await this.address()
 
   // Get block hash by height
   if (top && !isNaN(top)) {
@@ -163,7 +164,7 @@ async function contractCall (source, address, name, args = [], options = {}) {
   const opt = R.merge(this.Ae.defaults, options)
 
   const tx = await this.contractCallTx(R.merge(opt, {
-    callerId: await this.address(opt),
+    callerId: await this.address(),
     contractId: address,
     callData: await this.contractEncodeCall(source, name, args)
   }))
@@ -208,7 +209,7 @@ async function contractCall (source, address, name, args = [], options = {}) {
 async function contractDeploy (code, source, initState = [], options = {}) {
   const opt = R.merge(this.Ae.defaults, options)
   const callData = await this.contractEncodeCall(source, 'init', initState)
-  const ownerId = await this.address(opt)
+  const ownerId = await this.address()
 
   const { tx, contractId } = await this.contractCreateTx(R.merge(opt, {
     callData,
@@ -288,7 +289,7 @@ async function contractCompile (source, options = {}) {
  * const client = await ContractWithAe({ url, internalUrl, compilerUrl, keypair, ... })
  *
  */
-export const Contract = Ae.compose(ContractBase, {
+export const Contract = Ae.compose(ContractBase, ContractACI, {
   methods: {
     contractCompile,
     contractCallStatic,