+++ es/ae/universal.js
@@ -26,11 +26,11 @@ import Ae from './'
 import Account from '../account/memory'
 import Chain from '../chain/node'
 import Aens from './aens'
-import Contract from './contract'
 import Transaction from '../tx/tx'
 import Oracle from './oracle'
 import Selector from '../account/selector'
 import Accounts from '../accounts'
+import Contract from './contract'
 
 /**
  * Universal Stamp